requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-nested-inline==0.4.6
exceptiongroup==1.1.2
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pytest==7.4.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
